Split row and column labels into nouns and prepositional phrases

This function is similar to split_pref_suff() in that it returns a list. However, this function's list is more detailed than split_pref_suff(). The return value from this function is a list with the first named item being the prefix (with the name noun) followed by objects of prepositional phrases (with names being prepositions that precede the objects).

Usage

split_labels(
  labels,
  notation = RCLabels::bracket_notation,
  prepositions = RCLabels::prepositions
)

Arguments

labels

The row and column labels from which prepositional phrases are to be extracted.

notation

The notation object that describes the labels. Default is RCLabels::bracket_notation.

prepositions

A vector of strings to be treated as prepositions. Note that a space is appended to each word internally, so, e.g., "to" becomes "to ". Default is RCLabels::prepositions.

Value

A list of lists with items named noun and pp.

Details

Unlike split_pref_suff(), it does not make sense to have a transpose argument on split_labels(). Labels may not have the same structure, e.g., they may have different prepositions.

Examples

split_labels(c("a [of b in c]", "d [of e into f]"),
             notation = bracket_notation)
#> [[1]]
#> noun   of   in 
#>  "a"  "b"  "c" 
#> 
#> [[2]]
#> noun   of into 
#>  "d"  "e"  "f" 
#>